bin/ci/requirements.txt
PyYAML==6.*
filecache==0.81
flake8==7.*
gql==3.*
kaleido==0.2.*
pandas==2.*
plotly==5.*
python-dateutil==2.*
python-gitlab==5.*
rich==14.1.*
ruamel.yaml.clib==0.2.*
ruamel.yaml==0.17.*
tabulate==0.9.*
